compiles with skinning app
Former-commit-id: 18c9719d8e76b3076a0f547bac5b237736f19cbb

+++ include/igl/canonical_quaternions.cpp
@@ -2,11 +2,11 @@
 
 template <> IGL_INLINE const float igl::CANONICAL_VIEW_QUAT<float>(int i, int j)
 {
-  return igl::CANONICAL_VIEW_QUAT_F[i][j];
+  return (float)igl::CANONICAL_VIEW_QUAT_F[i][j];
 }
 template <> IGL_INLINE const double igl::CANONICAL_VIEW_QUAT<double>(int i, int j)
 {
-  return igl::CANONICAL_VIEW_QUAT_D[i][j];
+  return (double)igl::CANONICAL_VIEW_QUAT_D[i][j];
 }
 
 #ifndef IGL_HEADER_ONLY
